The simpleflush routine in stdio.d, called when printing a string or net, used to return before writing if it detected the interrupt flag had been set, but I changed that on Oct 12 when reworking interrupts.  Now I restore that, because it had the effect that user interrupts would no longer interrupt printouts.
I also take the opportunity have the routine to loop, writing the rest of the
buffer, if a write() call writes fewer bytes than requested.

Also, the interrupt handler will no longer print a newline character, because it can come out
in the middle of other output, which doesn't look good.


git-svn-id: svn://svn.macaulay2.com/Macaulay2/release-branches/1.3@9542 a8012f80-5f13-0410-8dc0-eb77bd940c0d